Add files via upload
feat: Restructured project architecture and implemented modular initialization system

Significant changes have been made to the project architecture to enhance its flexibility and support future extensions. Key modifications include:

1. **Modular architecture:** The project has been segmented into modules for better organization and code management, improving readability and maintainability.

2. **Initialization system:** Introduced a new initialization system to obtain essential information required to initialize the message generator engine. This ensures a more efficient and straightforward initialization process, reducing dependencies between different components.

3. **Performance improvements:** Code optimizations have been carried out to enhance the performance and efficiency of the message generator engine.

These changes aim to improve the overall project architecture and prepare it for future expansions. Details of the changes can be found in the respective project files and documentation.

+++ main.py
@@ -2,34 +2,34 @@ import discord
 from CuteON import Read_
 from sends import main
 from time import time
+from ChatBot import initBot
 
-CONFIG = Read_.readAll("Config.sws")
+BOT = initBot.init("Config.sws")
 
 class MyClient(discord.Client):
     async def on_ready(self):
         start = time()
         print("[Loading some bot...]")
-        print("[Token] - " + CONFIG["Token"])
-        print("[GitHub] - " + CONFIG["GitHub"])
-        print("[PlayIn] - " + CONFIG["PlayIn"])
+        print("[Token] - " + BOT["Token"])
+        print("[PlayIn] - " + BOT["About"]["PlayIn"])
         print('Logged on as', self.user)
         await client.change_presence(status=discord.Status.do_not_disturb, activity=discord.Game("Загрузка систем..."))
-        await client.change_presence(status=discord.Status.online, activity=discord.Game(CONFIG["PlayIn"] ))
+        await client.change_presence(status=discord.Status.online, activity=discord.Game(BOT["About"]["PlayIn"] ))
         print('[Time logged is] - ' + str(time() - start))
     async def on_message(self, message):
-
-        if message.content == '.cls':
-            channel = message.channel
-            messages = channel.history()
-            await message.channel.send("Конечно, сейчас очищю!")
-            async for message in messages:
-                await message.delete()
-
 
         # don't respond to ourselves
         if message.author == self.user:
             return
         if message.content == message.content:
+            if message.content[0] == BOT["comand_prefix"]:
+                if message.content == '.cls':
+                    channel = message.channel
+                    messages = channel.history()
+                    await message.channel.send("Конечно, сейчас очищю!")
+                    async for message in messages:
+                        await message.delete()
+
             #try:
                 mes = (message.content).lower()
                 mes_list = mes.split()
@@ -40,5 +40,5 @@ class MyClient(discord.Client):
 intents = discord.Intents.all()
 client = MyClient(intents=intents)
 
-print(CONFIG["Token"])
-client.run(str(CONFIG["Token"]))
+print(BOT["Token"])
+client.run(str(BOT["Token"]))
